+++ CHANGELOG.md
@@ -1,1 +1,9 @@
 # Change Log
+
+## 0.0.2
+
+Add support for syncing Project, Open Files, Directory
+
+## 0.0.1
+
+Initial release - support for syncing files on save.